node.js backend refactor branch
