+++ views/index.ejs
@@ -18,8 +18,7 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500&display=swap" rel="stylesheet"> 
     <link rel='stylesheet' href='/stylesheets/style.css' />    
     <script src="/javascripts/jquery-3.1.1.min.js"></script>
-    <script src='/javascripts/locationmap.js' /></script>
-    <script type="module" src='/javascripts/covidtracker.js' /></script>
+    <script src='/javascripts/covidtracker.js' /></script>
   </head>
   <body>
     <div id="header">
